backend/requirements.txt
# Web framework
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Settings & validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Auth
PyJWT==2.10.1
bcrypt==4.2.1
slowapi==0.1.9

# Payments
stripe==11.4.1

# Email
resend==2.5.1

# Dev / test
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
psycopg2-binary==2.9.10
ruff==0.8.6
mypy==1.14.1
